Move request utils to their own package
