+++ app/src/main/java/com/justsoft/redditshareinterceptor/processors/UnknownContentUrlPostProcessor.kt
@@ -5,7 +5,7 @@ import com.justsoft.redditshareinterceptor.model.RedditPost
 import com.justsoft.redditshareinterceptor.model.media.MediaContentType
 import com.justsoft.redditshareinterceptor.model.media.MediaContentType.TEXT
 import com.justsoft.redditshareinterceptor.model.media.MediaDownloadObject
-import com.justsoft.redditshareinterceptor.util.RequestHelper
+import com.justsoft.redditshareinterceptor.util.request.RequestHelper
 
 class UnknownContentUrlPostProcessor: PostProcessor {
     override fun isProcessorSuitableForPost(redditPost: RedditPost): Boolean = true
