+++ app/src/main/java/com/justsoft/redditshareinterceptor/util/request/VolleyRequestHelper.kt
@@ -1,4 +1,4 @@
-package com.justsoft.redditshareinterceptor.util
+package com.justsoft.redditshareinterceptor.util.request
 
 import com.android.volley.DefaultRetryPolicy
 import com.android.volley.RequestQueue
