+++ app/src/main/java/com/justsoft/redditshareinterceptor/websitehandlers/RedditUrlHandler.kt
@@ -12,8 +12,8 @@ import com.justsoft.redditshareinterceptor.model.media.MediaDownloadInfo
 import com.justsoft.redditshareinterceptor.model.media.MediaDownloadObject
 import com.justsoft.redditshareinterceptor.processors.*
 import com.justsoft.redditshareinterceptor.util.FirebaseAnalyticsHelper
-import com.justsoft.redditshareinterceptor.util.RequestHelper
 import com.justsoft.redditshareinterceptor.util.Stopwatch
+import com.justsoft.redditshareinterceptor.util.request.RequestHelper
 import java.util.regex.Pattern
 
 class RedditUrlHandler : UrlHandler {
