+++ app/src/test/java/com/justsoft/redditshareinterceptor/util/TestRequestHelperTest.kt
@@ -1,5 +1,6 @@
 package com.justsoft.redditshareinterceptor.util
 
+import com.justsoft.redditshareinterceptor.util.request.TestRequestHelper
 import org.junit.Assert.*
 import org.junit.Before
 import org.junit.Ignore
